feat: ViewType을 활용해 MainList와 Bottom 분리
