+++ app/src/main/java/com/rsupport/mobile1/test/presentation/main/ui/list/viewholder/MainBottomViewHolder.kt
@@ -1,0 +1,13 @@
+package com.rsupport.mobile1.test.presentation.main.ui.list.viewholder
+
+import androidx.recyclerview.widget.RecyclerView
+import com.rsupport.mobile1.test.databinding.ItemMainBottomBinding
+
+class MainBottomViewHolder(
+    private val binding: ItemMainBottomBinding
+) : RecyclerView.ViewHolder(binding.root) {
+
+    fun bind() {
+
+    }
+}
