+++ app/src/main/res/layout/activity_main.xml
@@ -10,96 +10,16 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@color/color_f4f4f4"
-        android:gravity="center">
+        android:background="@color/color_f4f4f4">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_main_list"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:spanCount="3"
             tools:itemCount="9"
-            tools:listitem="@layout/item_image_imformation" />
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="34dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/rv_main_list">
-
-            <EditText
-                android:id="@+id/et_page_number"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:background="@drawable/bg_stroke_080808"
-                android:paddingHorizontal="14dp"
-                android:paddingVertical="7dp"
-                android:text="@string/main_page_number"
-                android:textColor="@color/black"
-                android:textSize="14sp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toStartOf="@id/tv_of"
-                app:layout_constraintHorizontal_chainStyle="packed"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <TextView
-                android:id="@+id/tv_of"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="5dp"
-                android:text="@string/main_of"
-                android:textColor="@color/color_080808"
-                android:textSize="14sp"
-                app:layout_constraintBottom_toBottomOf="@id/et_page_number"
-                app:layout_constraintEnd_toStartOf="@id/tv_total_page_number"
-                app:layout_constraintStart_toEndOf="@id/et_page_number"
-                app:layout_constraintTop_toTopOf="@id/et_page_number" />
-
-            <TextView
-                android:id="@+id/tv_total_page_number"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="5dp"
-                android:text="@string/main_total_page_number"
-                android:textColor="@color/color_080808"
-                android:textSize="14sp"
-                app:layout_constraintBottom_toBottomOf="@id/et_page_number"
-                app:layout_constraintEnd_toStartOf="@id/btn_next"
-                app:layout_constraintStart_toEndOf="@id/tv_of"
-                app:layout_constraintTop_toTopOf="@id/et_page_number" />
-
-            <com.google.android.material.button.MaterialButton
-                android:id="@+id/btn_next"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_marginStart="30dp"
-                android:backgroundTint="@color/bg_white_to_black"
-                android:insetTop="0dp"
-                android:insetBottom="0dp"
-                android:paddingHorizontal="50dp"
-                android:paddingVertical="15dp"
-                android:stateListAnimator="@null"
-                android:text="@string/main_next"
-                android:textColor="@color/bg_black_to_white"
-                android:textSize="14sp"
-                app:icon="@drawable/ic_next_20dp"
-                app:iconGravity="textEnd"
-                app:iconPadding="0dp"
-                app:iconTint="@color/bg_black_to_white"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toEndOf="@id/tv_total_page_number"
-                app:layout_constraintTop_toTopOf="parent"
-                app:rippleColor="@android:color/transparent"
-                app:strokeColor="@color/black"
-                app:strokeWidth="1dp" />
-        </androidx.constraintlayout.widget.ConstraintLayout>
+            tools:listitem="@layout/item_main_list" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
